handle data step by step when completing requests
This reduces the risk of handling packages prompting move synchronous ubus
action.
More importantly, there's no nesting of ubus_handle_data in
ubus_complete_request which avoids ending up in a situation in which only
the innermost call is cancelled and the control flow never returns to any
of the outer _complete_requests (disabling the event loop and the handling
of invoke/unsubscribe/notify due to stack_depth being >0).

Signed-off-by: Leon M. Busch-George <leon@georgemail.eu>